+++ frontend/lib/main.dart
@@ -1,7 +1,7 @@
 // lib/main.dart
 
 import 'package:flutter/material.dart';
-import 'pages/login_page.dart'; // Import the LoginPage
+import 'presentation/pages/login_page.dart'; // Import the LoginPage
 
 void main() {
   runApp(MyApp());
@@ -15,7 +15,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Demo',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        primarySwatch: Colors.indigo,
       ),
       home: LoginPage(), // Set LoginPage as the home page
     );
