+++ frontend/lib/presentation/pages/home_page.dart
@@ -1,5 +1,3 @@
-// lib/pages/home_page.dart
-
 import 'package:flutter/material.dart';
 
 class HomePage extends StatelessWidget {
